==> 01_code/04_Wearables_PrePost_Report.Rmd <==
---
title: "Wearables Pre-Post Data Report"
author: "Laura Graham"
date: "`r Sys.Date()`"
output: 
  github_document:
    toc: true
    toc_depth: 2
---

Two analyses are used to assess the impact of DCI on physical wellbeing as measured by average daily steps, total minutes of moderate to vigorous activity (MVPA), and hours of sleep.

Analysis 1 uses a paired design (e.g. paired t-test). While this approach is simple, easy to interpret, and controls for individual differences because each person is compared to themselves, it only uses summary data per individual (e.g., mean or total steps), so it discards information about daily fluctuations. 

Analysis 2 uses a repeated measures model using day as the unit of analysis with a repeated effect for each individual and a fixed effect for the time period (Pre vs. Post). A repeated measures model makes use of all daily data, increasing statistical power, and can account for day-to-day variability, time trends, and other covariates. It's also more flexible, can handle missing days and unequal numbers of days per person. 

```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = TRUE)

library(purrr)
library(tidyverse)
library(jsonlite)
library(ggplot2)
library(dplyr)
library(tidyquant)
library(procs)
library(gtsummary)
library(readxl)
library(gridExtra)
library(lme4)
library(lsmeans)

theme_gtsummary_compact()

# Load Current Step Data
  load("~/Library/CloudStorage/GoogleDrive-lagraham@stanford.edu/Shared drives/Secure: DCI Research/Analysis_Graham/Wearables Data/Daily_Steps.RData")

# Load Current HR Data
  load("~/Library/CloudStorage/GoogleDrive-lagraham@stanford.edu/Shared drives/Secure: DCI Research/Analysis_Graham/Wearables Data/Daily_MVPA.RData")

# Load Current Sleep Data
  load("~/Library/CloudStorage/GoogleDrive-lagraham@stanford.edu/Shared drives/Secure: DCI Research/Analysis_Graham/Wearables Data/Daily_Sleep.RData")

# Merge all Wearables
  daily <- merge(daily_steps, daily_mvpa, by=c("ID", "date"), all.x=T, all.y=T)
  daily <- merge(daily, daily_sleep, by=c("ID", "date"), all.x=T, all.y=T)
  
  daily <- daily[,c("ID", "date", "Total_Steps", "Total_MVPA", "Total_Sleep", "HR_Avg", "HR_Max", "HR_n", "Moderate", "Vigorous", "Sleep_Start", "Sleep_End")]

  # Participant level with day daily
    participant <- daily %>%
      dplyr::group_by(ID) %>%
      dplyr::summarize(Avg_Daily_Steps = mean(Total_Steps, na.rm=T), 
                       Avg_MVPA = mean(Total_MVPA, na.rm=T),
                       Avg_Sleep = mean(Total_Sleep, na.rm=T),
                       start_date = min(date), last_date = max(date)) 
  
    # Pre-Post Determinimation
    daily$Pre.Post <- ifelse(daily$date < as.Date('2024-09-15'), "Pre", "Post")
    daily$Pre.Post <- factor(daily$Pre.Post, levels=c("Pre", "Post"))
      table(daily$Pre.Post)
    daily$Times <- ifelse(daily$date < as.Date('2024-05-16'), "1. May 15 or earlier",
                          ifelse(daily$date <= as.Date('2024-06-28'), "2. Pre Intro to DCI (06/28/2024)",
                                 ifelse(daily$date <= as.Date('2024-08-01'), "3. Pre Virtual Orientation (08/01/2024)",
                                        ifelse(daily$date < as.Date('2024-09-15'), "4. Pre In-Person Orientation (09/15/2024)", "5. Post DCI Orientation"))))
      table(daily$Times)

# Demographics
  Demo <- read_excel("~/Library/CloudStorage/GoogleDrive-lagraham@stanford.edu/Shared drives/Secure: DCI Research/Analysis_Graham/Wearables Data/2024 Master List.xlsx", sheet = "Master")

# Survey Data
  load("~/Library/CloudStorage/GoogleDrive-lagraham@stanford.edu/Shared drives/Secure: DCI Research/Analysis_Graham/Survey Data/Survey_2024_07092024.RData")
  survey <- survey2024[,c("id_number2", "Q156_1", "Q127", "Q129", "Q153")]
  
# TruDiagnostics
  Trudiagnostics <- read.csv("~/Library/CloudStorage/GoogleDrive-lagraham@stanford.edu/Shared drives/Secure: DCI Research/Analysis_Graham/TruDiagnostics Data/PopulationData_11192024.csv", header=TRUE)
  Trudiagnostics <- Trudiagnostics[,c("Patient.ID", "Telomere.Values", "OMICmAge", "Decimal.Chronological.Age")]

# Combine all data
  daily2 <- merge(daily, Demo, by = "ID", all.x = T)
  daily2 <- merge(daily2, survey, by.x = "Old ID", by.y = "id_number2", all.x = T)
  daily2 <- merge(daily2, Trudiagnostics, by.x = "TruDiagnosticID", by.y = "Patient.ID", all.x = T)

  # Subsetting Data
    daily2 <- subset(daily2, date > as.Date("2024-05-16"))
    daily2 <- subset(daily2, date < as.Date("2025-06-01"))
    
    names(daily2)
 
  daily2 <- daily2[,c("ID", "date", "Total_Steps", "Total_MVPA", "Total_Sleep", "HR_Avg", "HR_Max", "HR_n", "Moderate", "Vigorous", "Sleep_Start", "Sleep_End", "Pre.Post", "Times",
                  "Cohort", "gender", "Surv_year", "birth_year", "Age", "education", 
                  "Usborn", "married",  "children_living2", "selfemployed", "work_years", "rate_health", "Big5_Agreeable", "Big5_Conscientious",
                  "Big5_Extravert", "Big5_Neurotic", "Big5_Open", "UCLA_Loneliness3", "UCLA_Loneliness_Rev", "Need_for_Cognition", "Ryff_Purpose", "Q156_1",  
                  "Q127", "Q129", "Q153", "Telomere.Values", "OMICmAge", "Decimal.Chronological.Age")]

# Cleaning Labels
  daily2 <- labelled::copy_labels(from = survey, to = daily2)

  labelled::var_label(daily2$Pre.Post) <- "Pre/Post DCI"
  labelled::var_label(daily2$gender) <- "Gender"
  labelled::var_label(daily2$education) <- "Education"
  labelled::var_label(daily2$Usborn) <- "Born in US"
  labelled::var_label(daily2$married) <- "Marital Status"
  labelled::var_label(daily2$children_living2) <- "Number of Children"
  labelled::var_label(daily2$selfemployed) <- "Self-Employed"
    daily2$rate_health <- factor(daily2$rate_health, levels = c("Excellent", "Very Good", "Good", "Fair", "Poor"))
  labelled::var_label(daily2$rate_health) <- "Self-Reported Health"
  daily2$Q156_1<- plyr::revalue(as.character(daily2$Q156_1), 
                                     c("5" = "Very Satisfied", 
                                       "4" = "Satisfied", 
                                       "3" = "Neither Dissatisfied nor Satisfied", 
                                       "2" = "Dissatisfied", 
                                       "1" = "Very Dissatisfied"))
  labelled::var_label(daily2$Q156_1) <- "How satisfied are you with your health?"
  daily2 <- labelled::copy_labels(from = survey, to = daily2)
  

```

# Distribution of Daily Wearable Measures

Unit of analysis is the day.

```{r, warning=FALSE, echo=FALSE, message = FALSE, fig.height=8}

p1 <- ggplot(daily2, aes(x = Total_Steps, y = ..density..)) +
  geom_histogram(fill = "cornsilk", colour = "grey60", size = .2) +
  geom_density(colour = "red") + theme_bw() +  
  ggtitle("Average Daily Step Counts Across All Time for Participants")+
  labs(x = "Average Daily Steps", y = "Density")

p2 <- ggplot(daily2, aes(x = Total_MVPA, y = ..density..)) +
  geom_histogram(fill = "cornsilk", colour = "grey60", size = .2) +
  geom_density(colour = "red") + theme_bw() +  
  ggtitle("Average Daily MVPA Across All Time for Participants")+
  labs(x = "Average Daily MVPA", y = "Density")

p3 <- ggplot(daily2, aes(x = Total_Sleep, y = ..density..)) +
  geom_histogram(fill = "cornsilk", colour = "grey60", size = .2) +
  geom_density(colour = "red") + theme_bw() +  
  ggtitle("Average Daily Sleep Hours Across All Time for Participants")+
  labs(x = "Average Daily Sleep Hours", y = "Density")

grid.arrange(p1, p2, p3, nrow = 3)

```

# 1. Participant-Level Analysis
## Participant Characteristics

Unit of analysis is the person.

```{r, warning=FALSE, echo=FALSE, message = FALSE, results='markup'}

# Overall Averages
  participant <- daily2 %>%
    dplyr::group_by(ID) %>%
    dplyr::summarize(Avg_Daily_Steps = mean(Total_Steps, na.rm=T), 
                     Avg_Daily_MVPA = mean(Total_MVPA, na.rm=T),
                     Avg_Daily_Sleep = mean(Total_Sleep, na.rm=T),
                     Med_Daily_Steps = median(Total_Steps, na.rm=T), 
                     Med_Daily_MVPA = median(Total_MVPA, na.rm=T),
                     Med_Daily_Sleep = median(Total_Sleep, na.rm=T),
                     
                     start_date = min(date), last_date = max(date)) 
  
  participant$time <- participant$last_date-participant$start_date

# Limit to those with at least one pre/post assessment
  Pre <- subset(daily2, Pre.Post == "Pre") %>% group_by(ID) %>%
      summarize(Pre_Steps = mean(Total_Steps, na.rm=T), 
                Pre_MVPA = mean(Total_MVPA, na.rm=T),
                Pre_Sleep = mean(Total_Sleep, na.rm=T))
  
  Post <- subset(daily2, Pre.Post == "Post") %>% group_by(ID) %>%
    summarize(Post_Steps = mean(Total_Steps, na.rm=T), 
              Post_MVPA = mean(Total_MVPA, na.rm=T),
              Post_Sleep = mean(Total_Sleep, na.rm=T))
  
  Pre.Post <- merge(Pre, Post, by=("ID"))
  participant <- merge(Pre.Post, participant, by=("ID"))

# Merge in demographics
  Demo2 <- daily2[,c("ID", "Cohort", "gender", "Surv_year", "Age", "education", "Usborn", "married", "children_living2",
                     "selfemployed", "work_years", "rate_health")]
  Demo2 <- distinct(Demo2)
  participant <- merge(participant, Demo2, by="ID", all.x=T)
  
# Variable Levels
  participant <- labelled::copy_labels(from = survey, to = participant)
  
  labelled::var_label(participant$Avg_Daily_Steps) <- "Average Daily Steps"
  labelled::var_label(participant$Avg_Daily_MVPA) <- "Average Daily Minutes of Moderate to Vigorous Physcial Activity (MVPA)"
  labelled::var_label(participant$Avg_Daily_Sleep) <- "Average Daily Sleep Hours"
  labelled::var_label(participant$start_date) <- "Earliest Recorded Date"
  labelled::var_label(participant$last_date) <- "Last Recorded Date"
  labelled::var_label(participant$time) <- "Days of Observation"
  labelled::var_label(participant$gender) <- "Sex"
  labelled::var_label(participant$Age) <- "Age, years"
  labelled::var_label(participant$Usborn) <- "US born"
  labelled::var_label(participant$married) <- "Marital Status"
  labelled::var_label(participant$children_living2) <- "Number of Children"
  labelled::var_label(participant$selfemployed) <- "Self-Employed"
  labelled::var_label(participant$work_years) <- "Working Years"
  labelled::var_label(participant$rate_health) <- "Self-Rated Health at the Start of DCI"
  
tbl_summary(participant, include = c("Avg_Daily_Steps", "Avg_Daily_MVPA", "Avg_Daily_Sleep", "start_date", "last_date", "time", "gender", "Age", "Usborn", "married", "children_living2", "selfemployed", "work_years", "rate_health")) %>% 
  as_gt() %>% gt::as_raw_html()

```

## Participant Characteristics Associated with Wearable Metrics

```{r, echo=FALSE, message = FALSE, warning=FALSE}

participant$Age_Category <- gtools::quantcut(participant$Age, q = 4, na.rm = TRUE)
participant$Work_Years_Category <- gtools::quantcut(participant$work_years, q = 4, na.rm = TRUE)

t1 <- tbl_continuous(data=participant, Avg_Daily_Steps, 
               include = c("gender", "Age_Category", "Usborn", "married", "children_living2", "selfemployed", "Work_Years_Category", "rate_health")) %>% add_p() %>% bold_p() 
t2 <- tbl_continuous(data=participant, Avg_Daily_MVPA, 
               include = c("gender", "Age_Category", "Usborn", "married", "children_living2", "selfemployed", "Work_Years_Category", "rate_health")) %>% add_p() %>% bold_p()
t3 <- tbl_continuous(data=participant, Avg_Daily_Sleep, 
               include = c("gender", "Age_Category", "Usborn", "married", "children_living2", "selfemployed", "Work_Years_Category", "rate_health")) %>% add_p() %>% bold_p()

tbl_merge(
  tbls = list(t1, t2, t3),
  tab_spanner = c("**Average Daily Steps**", "**Average Daily MVPA**", "**Average Daily Sleep Hours**")
) %>%
  as_gt() %>% gt::tab_source_note(gt::md("*Data as of June 1, 2025*")) %>% gt::as_raw_html()

```

## Change in Wearable Metrics by Study Time Period (Pre vs. Post)

```{r, echo=FALSE, message = FALSE, warning=FALSE, fig.height=3.5}

Pre <- subset(daily2, Pre.Post == "Pre") %>% group_by(ID) %>%
    summarize(Pre_Steps = mean(Total_Steps, na.rm=T), 
              Pre_MVPA = mean(Total_MVPA, na.rm=T),
              Pre_Sleep = mean(Total_Sleep, na.rm=T))

Post <- subset(daily2, Pre.Post == "Post") %>% group_by(ID) %>%
  summarize(Post_Steps = mean(Total_Steps, na.rm=T), 
            Post_MVPA = mean(Total_MVPA, na.rm=T),
            Post_Sleep = mean(Total_Sleep, na.rm=T))

Pre.Post <- merge(Pre, Post, by=("ID"))

ggpubr::ggpaired(Pre.Post[complete.cases(Pre.Post[ , c('Pre_Steps', 'Post_Steps')]), ],
                 cond1 = "Pre_Steps", cond2 = "Post_Steps", fill="condition", line.color = "gray", line.size = 0.2,
                palette = "jama") + ggtitle("Average Daily Steps") + ylab('Mean') + xlab("") + labs(fill="Time") +
  theme_bw() + ggpubr::stat_compare_means(method="wilcox.test", paired=TRUE, hjust=-.2)

ggpubr::ggpaired(Pre.Post[complete.cases(Pre.Post[ , c('Pre_MVPA', 'Post_MVPA')]), ],
                 cond1 = "Pre_MVPA", cond2 = "Post_MVPA", fill="condition", line.color = "gray", line.size = 0.2,
                 palette = "jama") + ggtitle("Average Daily MVPA") + ylab('Mean') + xlab("") + labs(fill="Time") +
  theme_bw() + ggpubr::stat_compare_means(method="wilcox.test", paired=TRUE, hjust=-.2)

ggpubr::ggpaired(Pre.Post[complete.cases(Pre.Post[ , c('Pre_Sleep', 'Post_Sleep')]), ],
                 cond1 = "Pre_Sleep", cond2 = "Post_Sleep", fill="condition", line.color = "gray", line.size = 0.2,
                 palette = "jama") + ggtitle("Average Daily Sleep Hours") + ylab('Mean') + xlab("") + labs(fill="Time") +
  theme_bw() + ggpubr::stat_compare_means(method="wilcox.test", paired=TRUE, hjust=-.2)

```

# 2. Day-Level Analysis
## Unadjusted

```{r, echo=FALSE, message = FALSE, warning=FALSE}

daily3 <- daily2[daily2$ID %in% participant$ID, ] 

tbl_summary(data = daily3, by=Pre.Post, include=c(Total_Steps, Total_MVPA, Total_Sleep), 
            type = all_continuous() ~ "continuous2",
            statistic = all_continuous() ~ c(
              "{median} ({p25}, {p75})",
              "{min}, {max}"                  ),
            label = c(Total_Steps ~ "Daily Steps", 
                      Total_MVPA ~ "Daily Minutes Moderate to Vigorous Activity",
                      Total_Sleep ~ "Daily Hours of Sleep")) %>% 
  add_overall() %>% add_p() %>% 
  as_gt() %>% gt::as_raw_html()

```

## Adjusted for age, sex, and self-reported health

ID is included as a repeated effect to account for clustering on days within a participant.

```{r, echo=FALSE, message = FALSE, warning=FALSE}

t1 <- tbl_regression(glmer(data = daily3, 
                          Total_Steps ~ Pre.Post + gender + Age + rate_health + (1|ID))) %>% 
                          add_glance_source_note() %>% add_global_p() %>%
                          modify_header(label = "**Adjusted Model**")

t2 <- tbl_regression(glmer(data = daily3, 
                          Total_MVPA ~ Pre.Post + gender + Age + rate_health + (1|ID))) %>% 
                          add_glance_source_note() %>% add_global_p() %>%
                          modify_header(label = "**Adjusted Model**")

t3 <- tbl_regression(glmer(data = daily3, 
                          Total_Sleep ~ Pre.Post + gender + Age + rate_health + (1|ID))) %>% 
                          add_glance_source_note() %>% add_global_p() %>%
                          modify_header(label = "**Adjusted Model**")

tbl_merge(tbls = list(t1, t2, t3),
  tab_spanner = c("**Change in Daily Steps**", "**Change in Daily MVPA**", "**Change in Daily Sleep Hours**")
) %>%
  as_gt() %>% gt::tab_source_note(gt::md("*Data as of June 1, 2025*")) %>% gt::as_raw_html()

```

# Exploring Interactions

## Difference in Effect of DCI on Step Count by Gender

```{r, echo=FALSE, message = FALSE, warning=FALSE, fig.height=3}

daily3$Pre.Post <- relevel(daily3$Pre.Post, ref="Pre")
t1 <- tbl_regression(glmer(data = subset(daily3, gender == "Male"), Total_Steps ~ Pre.Post + (1|ID), family = poisson(link="identity")),
                     label = list(Pre.Post ~ "Difference in Steps Pre/Post Among Males")) %>% bold_p()
t2 <- tbl_regression(glmer(data = subset(daily3, gender == "Female"), Total_Steps ~ Pre.Post + (1|ID), family = poisson(link="identity")),
                     label = list(Pre.Post ~ "Difference in Steps Pre/Post Among Females")) %>% bold_p()
tbl_stack(tbls = list(t1, t2)) %>% as_gt() %>% gt::tab_source_note(gt::md("*Data as of June 1, 2025*")) %>% gt::as_raw_html()

# Diff-in-diff Plot
  mod.male <- glmer(data = subset(daily3, gender == "Male"), Total_Steps ~ Pre.Post + (1|ID))
  mod.female <- glmer(data = subset(daily3, gender == "Female"), Total_Steps ~ Pre.Post + (1|ID))
  
  pre.male <- as.data.frame(lsmeans(mod.male, "Pre.Post")[1])
  post.male <- as.data.frame(lsmeans(mod.male, "Pre.Post")[2])
  pre.female <- as.data.frame(lsmeans(mod.female, "Pre.Post")[1])
  post.female <- as.data.frame(lsmeans(mod.female, "Pre.Post")[2])
  
  male <- rbind(pre.male, post.male)
  male$Gender <- "Male"
  female <- rbind(pre.female, post.female)
  female$Gender <- "Female"
  
  plot_data <- rbind(male, female)

  # Plot
  ggplot(plot_data, aes(x = Pre.Post, y = lsmean, color = Gender, label=round(lsmean, digits = 1))) +
    geom_pointrange(aes(ymin = SE, ymax = SE), size = 1) +
    geom_line(aes(group = Gender)) +
    geom_text(hjust=1, vjust=-1) +
    theme_bw() + coord_cartesian(ylim=c(10000, 15000)) +
    labs(title="Interaction of Sex and Time", y = "Average Daily Steps", x = "")

# Age-Adjusted
daily3$Pre.Post <- relevel(daily3$Pre.Post, ref="Post")
daily3$gender <- relevel(as.factor(daily3$gender), ref="Male")
tbl_regression(glmer(data = daily3, Total_Steps ~ Pre.Post:gender + Age + (1|ID), family = poisson(link="identity"))) %>%
  modify_header(label = "**Interaction with Adjustment**")  %>% add_global_p() %>% bold_p() %>% 
  as_gt() %>% gt::tab_source_note(gt::md("*Data as of June 1, 2025*")) %>% gt::as_raw_html()

```


## Difference in Effect of DCI on MVPA by Gender

```{r, echo=FALSE, message = FALSE, warning=FALSE, fig.height=3}

daily3$Pre.Post <- relevel(daily3$Pre.Post, ref="Pre")
t1 <- tbl_regression(glmer(data = subset(daily3, gender == "Male"), Total_MVPA ~ Pre.Post + (1|ID)),
                     label = list(Pre.Post ~ "Difference in MVPA Pre/Post Among Males")) 
t2 <- tbl_regression(glmer(data = subset(daily3, gender == "Female"), Total_MVPA ~ Pre.Post + (1|ID)),
                     label = list(Pre.Post ~ "Difference in MVPA Pre/Post Among Females")) 
tbl_stack(tbls = list(t1, t2)) %>% as_gt() %>% gt::tab_source_note(gt::md("*Data as of June 1, 2025*")) %>% gt::as_raw_html()

# Diff-in-diff Plot
  mod.male <- glmer(data = subset(daily3, gender == "Male"), Total_MVPA ~ Pre.Post + (1|ID))
  mod.female <- glmer(data = subset(daily3, gender == "Female"), Total_MVPA ~ Pre.Post + (1|ID))
  
  pre.male <- as.data.frame(lsmeans(mod.male, "Pre.Post")[1])
  post.male <- as.data.frame(lsmeans(mod.male, "Pre.Post")[2])
  pre.female <- as.data.frame(lsmeans(mod.female, "Pre.Post")[1])
  post.female <- as.data.frame(lsmeans(mod.female, "Pre.Post")[2])
  
  male <- rbind(pre.male, post.male)
  male$Gender <- "Male"
  female <- rbind(pre.female, post.female)
  female$Gender <- "Female"
  
  plot_data <- rbind(male, female)

  # Plot
  ggplot(plot_data, aes(x = Pre.Post, y = lsmean, color = Gender, label=round(lsmean, digits=1))) +
    geom_pointrange(aes(ymin = SE, ymax = SE), size = 1) +
    geom_line(aes(group = Gender)) +
    geom_text(hjust=0, vjust=-1) +
    theme_bw() + coord_cartesian(ylim=c(80,120)) +
    labs(title="Interaction of Sex and Time", y = "Average Daily MVPA", x = "")

# Age-Adjusted
daily3$Pre.Post <- relevel(daily3$Pre.Post, ref="Post")
daily3$gender <- relevel(as.factor(daily3$gender), ref="Male")
tbl_regression(glmer(data = daily3, Total_MVPA ~ Pre.Post:gender + Age + (1|ID))) %>%
  modify_header(label = "**Interaction with Adjustment**")  %>% add_global_p() %>%
  as_gt() %>% gt::tab_source_note(gt::md("*Data as of June 1, 2025*")) %>% gt::as_raw_html() 

```


## Difference in Effect of DCI on Sleep by Gender

```{r, echo=FALSE, message = FALSE, warning=FALSE, fig.height=3}

daily3$Pre.Post <- relevel(daily3$Pre.Post, ref="Pre")
t1 <- tbl_regression(glmer(data = subset(daily3, gender == "Male"), Total_Sleep ~ Pre.Post + (1|ID)),
                     label = list(Pre.Post ~ "Difference in Sleep Pre/Post Among Males")) 
t2 <- tbl_regression(glmer(data = subset(daily3, gender == "Female"), Total_Sleep ~ Pre.Post + (1|ID)),
                     label = list(Pre.Post ~ "Difference in Sleep Pre/Post Among Females")) 
tbl_stack(tbls = list(t1, t2)) %>% as_gt() %>% gt::tab_source_note(gt::md("*Data as of June 1, 2025*")) %>% gt::as_raw_html()

# Diff-in-diff Plot
  mod.male <- glmer(data = subset(daily3, gender == "Male"), Total_Sleep ~ Pre.Post + (1|ID))
  mod.female <- glmer(data = subset(daily3, gender == "Female"), Total_Sleep ~ Pre.Post + (1|ID))
  
  pre.male <- as.data.frame(lsmeans(mod.male, "Pre.Post")[1])
  post.male <- as.data.frame(lsmeans(mod.male, "Pre.Post")[2])
  pre.female <- as.data.frame(lsmeans(mod.female, "Pre.Post")[1])
  post.female <- as.data.frame(lsmeans(mod.female, "Pre.Post")[2])
  
  male <- rbind(pre.male, post.male)
  male$Gender <- "Male"
  female <- rbind(pre.female, post.female)
  female$Gender <- "Female"
  
  plot_data <- rbind(male, female)

  # Plot
  ggplot(plot_data, aes(x = Pre.Post, y = lsmean, color = Gender, label=round(lsmean, digits=2))) +
    geom_pointrange(aes(ymin = SE, ymax = SE), size = 1) +
    geom_line(aes(group = Gender)) +
    geom_text(hjust=-.5, vjust=-1) +
    theme_bw() + coord_cartesian(ylim=c(7,8)) +
    labs(title="Interaction of Sex and Time", y = "Average Daily Sleep Hours", x = "")

# Age-Adjusted
daily3$Pre.Post <- relevel(daily3$Pre.Post, ref="Post")
daily3$gender <- relevel(as.factor(daily3$gender), ref="Male")
tbl_regression(glmer(data = daily3, Total_Sleep ~ Pre.Post:gender + Age + (1|ID))) %>%
  modify_header(label = "**Interaction with Adjustment**")  %>% add_global_p() %>%
  as_gt() %>% gt::tab_source_note(gt::md("*Data as of June 1, 2025*")) %>% gt::as_raw_html()

```
